+++ timetracker/timetracker.csproj
@@ -75,6 +75,7 @@
     <Compile Include="Settings.cs" />
     <Compile Include="KimaiDatabase.cs" />
     <Compile Include="SqliteActivity.cs" />
+    <Compile Include="About.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\AboutResources.txt" />
@@ -97,6 +98,7 @@
     <AndroidResource Include="Resources\mipmap-xhdpi\ic_launcher.png" />
     <AndroidResource Include="Resources\mipmap-xxhdpi\ic_launcher.png" />
     <AndroidResource Include="Resources\mipmap-xxxhdpi\ic_launcher.png" />
+    <AndroidResource Include="Resources\layout\About.axml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Resources\drawable\" />
